pyproject.toml
[project]
name = "sudoku-solver"
version = "0.1.0"
description = "Experimental sudoku solver"
authors = [
    { name = "Andrei Sviridov", email = "mrlokans@gmail.com" }
]
dependencies = [
    "opencv-python>=4.9.0.80",
    "imutils>=0.5.4",
    "keras>=3.1.1",
    "tensorflow>=2.16.1",
    "pandas>=2.2.1",
    "scikit-learn>=1.4.1.post1",
    "scikit-image>=0.22.0",
    "pytesseract>=0.3.10",
    "python-telegram-bot>=21.0.1",
    "click>=8.1.7",
]
readme = "README.md"
requires-python = ">= 3.11,<3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.1.1",
    "mypy>=1.9.0",
    "tensorflow-datasets>=4.9.4",
    "jupyter>=1.0.0",
    "notebook>=7.1.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bot", "src/recognizer", "src/sudoku"]
